move metanome-execution script into metanome folder
