+++ server/routes/runMetanome.ts
@@ -5,7 +5,7 @@ import { rename } from "fs/promises";
 import { join } from "path";
 import { split } from "../utils/databaseUtils";
 import { pathSplit } from "../utils/files";
-import MetanomeAlgorithm from "./metanomeAlgorithm";
+import MetanomeAlgorithm from "../metanome/metanomeAlgorithm";
 
 export default function postRunMetanomeFDAlgorithmFunction(): RequestHandler {
   async function MetanomeFDAlgorithm(
